+++ common/src/main/java/traben/entity_texture_features/features/property_reading/properties/RandomProperties.java
@@ -34,6 +34,7 @@ public class RandomProperties {
                 VariantProperty::getPropertyOrNull,
                 CreativeProperty::getPropertyOrNull,
                 TeammateProperty::getPropertyOrNull,
+                ClientPlayerProperty::getPropertyOrNull,
 
                 //etf externals
                 HourProperty::getPropertyOrNull,